refactor(client): split recordings and settings pages into hooks and panels
Extract session review, account settings, and room settings into dedicated
hooks and section components with destructured query usage.
